requirements.txt
fastapi
uvicorn
groq
pymongo
python-dotenv
streamlit
requests
